backend/requirements.txt
fastapi
uvicorn
requests
Pillow
python-dotenv